Round out testing on read resource for test_one.
Introduce ProviderMeta to ReadResourceRequest, check it's being passed
in the tests.

Split serve tests into multiple files, they're getting impossible to
navigate.

Finish testing read resource for test_one resource, still need to test
it with test_two resource.
