+++ serve.go
@@ -320,6 +320,32 @@ func (s *server) ReadResource(ctx context.Context, req *tfprotov6.ReadResourceRe
 			Schema: resourceSchema,
 		},
 	}
+	if pm, ok := s.p.(ProviderWithProviderMeta); ok {
+		pmSchema, diags := pm.GetMetaSchema(ctx)
+		if diags != nil {
+			resp.Diagnostics = append(resp.Diagnostics, diags...)
+			if diagsHasErrors(resp.Diagnostics) {
+				return resp, nil
+			}
+		}
+		readReq.ProviderMeta = Config{
+			Schema: pmSchema,
+			Raw:    tftypes.NewValue(pmSchema.TerraformType(ctx), nil),
+		}
+
+		if req.ProviderMeta != nil {
+			pmValue, err := req.ProviderMeta.Unmarshal(pmSchema.TerraformType(ctx))
+			if err != nil {
+				resp.Diagnostics = append(resp.Diagnostics, &tfprotov6.Diagnostic{
+					Severity: tfprotov6.DiagnosticSeverityError,
+					Summary:  "Error parsing provider_meta",
+					Detail:   "There was an error parsing the provider_meta block. Please report this to the provider developer:\n\n" + err.Error(),
+				})
+				return resp, nil
+			}
+			readReq.ProviderMeta.Raw = pmValue
+		}
+	}
 	readResp := ReadResourceResponse{
 		State: State{
 			Schema: resourceSchema,
