+++ serve_resource_one_test.go
@@ -1,0 +1,104 @@
+package tfsdk
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/hashicorp/terraform-plugin-framework/schema"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
+	"github.com/hashicorp/terraform-plugin-go/tftypes"
+)
+
+type testServeResourceTypeOne struct{}
+
+func (rt testServeResourceTypeOne) GetSchema(_ context.Context) (schema.Schema, []*tfprotov6.Diagnostic) {
+	return schema.Schema{
+		Version: 1,
+		Attributes: map[string]schema.Attribute{
+			"name": {
+				Required: true,
+				Type:     types.StringType,
+			},
+			"favorite_colors": {
+				Optional: true,
+				Type:     types.ListType{ElemType: types.StringType},
+			},
+			"created_timestamp": {
+				Computed: true,
+				Type:     types.StringType,
+			},
+		},
+	}, nil
+}
+
+func (rt testServeResourceTypeOne) NewResource(p Provider) (Resource, []*tfprotov6.Diagnostic) {
+	provider, ok := p.(*testServeProvider)
+	if !ok {
+		prov, ok := p.(*testServeProviderWithMetaSchema)
+		if !ok {
+			panic(fmt.Sprintf("unexpected provider type %T", p))
+		}
+		provider = prov.testServeProvider
+	}
+	return testServeResourceOne{
+		provider: provider,
+	}, nil
+}
+
+var testServeResourceTypeOneSchema = &tfprotov6.Schema{
+	Version: 1,
+	Block: &tfprotov6.SchemaBlock{
+		Attributes: []*tfprotov6.SchemaAttribute{
+			{
+				Name:     "created_timestamp",
+				Computed: true,
+				Type:     tftypes.String,
+			},
+			{
+				Name:     "favorite_colors",
+				Optional: true,
+				Type:     tftypes.List{ElementType: tftypes.String},
+			},
+			{
+				Name:     "name",
+				Required: true,
+				Type:     tftypes.String,
+			},
+		},
+	},
+}
+
+var testServeResourceTypeOneType = tftypes.Object{
+	AttributeTypes: map[string]tftypes.Type{
+		"created_timestamp": tftypes.String,
+		"favorite_colors":   tftypes.List{ElementType: tftypes.String},
+		"name":              tftypes.String,
+	},
+}
+
+type testServeResourceOne struct {
+	provider *testServeProvider
+}
+
+func (r testServeResourceOne) Create(_ context.Context, _ CreateResourceRequest, _ *CreateResourceResponse) {
+	panic("not implemented") // TODO: Implement
+}
+
+func (r testServeResourceOne) Read(ctx context.Context, req ReadResourceRequest, resp *ReadResourceResponse) {
+	r.provider.readResourceCurrentStateValue = req.State.Raw
+	r.provider.readResourceCurrentStateSchema = req.State.Schema
+	r.provider.readResourceProviderMetaValue = req.ProviderMeta.Raw
+	r.provider.readResourceProviderMetaSchema = req.ProviderMeta.Schema
+	r.provider.readResourceCalledResourceType = "test_one"
+	r.provider.readResourceImpl(ctx, req, resp)
+}
+
+func (r testServeResourceOne) Update(_ context.Context, _ UpdateResourceRequest, _ *UpdateResourceResponse) {
+	panic("not implemented") // TODO: Implement
+}
+
+func (r testServeResourceOne) Delete(_ context.Context, _ DeleteResourceRequest, _ *DeleteResourceResponse) {
+	panic("not implemented") // TODO: Implement
+}
